+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # :package: 0.1.0
 
+- :gift: Added class `dev.orne.beans.converters.LocaleConverter`